# apps/backend/src/clock/clock.controller.ts
import { Controller, Post, Get, Body, UseGuards, Req } from '@nestjs/common';
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiBearerAuth,
} from '@nestjs/swagger';
import { ClockService } from './clock.service';
import {
  ClockInDto,
  ClockOutDto,
  ClockRecordResponseDto,
} from './dto/clock.dto';
import { JwtAuthGuard } from '../auth/jwt-auth.guard';

@ApiTags('clock')
@Controller('clock')
@UseGuards(JwtAuthGuard)
@ApiBearerAuth()
export class ClockController {
  constructor(private readonly clockService: ClockService) {}

  @Post('in')
  @ApiOperation({ summary: '出勤打刻', description: 'Register clock-in time' })
  @ApiResponse({
    status: 201,
    description: 'Clock-in recorded successfully',
    type: ClockRecordResponseDto,
  })
  @ApiResponse({ status: 401, description: 'Unauthorized' })
  async clockIn(
    @Req() req: any,
    @Body() clockInDto: ClockInDto,
  ): Promise<ClockRecordResponseDto> {
    const userId = req.user.userId;
    return this.clockService.clockIn(
      userId,
      clockInDto.location,
      clockInDto.deviceId,
    );
  }

  @Post('out')
  @ApiOperation({ summary: '退勤打刻', description: 'Register clock-out time' })
  @ApiResponse({
    status: 201,
    description: 'Clock-out recorded successfully',
    type: ClockRecordResponseDto,
  })
  @ApiResponse({ status: 401, description: 'Unauthorized' })
  async clockOut(
    @Req() req: any,
    @Body() clockOutDto: ClockOutDto,
  ): Promise<ClockRecordResponseDto> {
    const userId = req.user.userId;
    return this.clockService.clockOut(
      userId,
      clockOutDto.location,
      clockOutDto.deviceId,
    );
  }

  @Get('records')
  @ApiOperation({
    summary: '打刻記録一覧取得',
    description: 'Get all clock records for the authenticated user',
  })
  @ApiResponse({
    status: 200,
    description: 'List of clock records',
    type: [ClockRecordResponseDto],
  })
  @ApiResponse({ status: 401, description: 'Unauthorized' })
  async getRecords(@Req() req: any): Promise<ClockRecordResponseDto[]> {
    const userId = req.user.userId;
    return this.clockService.getRecords(userId);
  }
}